\name{ppiPre-internal}
\title{Internal ppiPre objects}
\alias{.initial}
\alias{CheckAnnotationPackage}
\alias{GetGOMap}
\alias{GetGOParents}
\alias{GetLatestCommonAncestor}
\alias{GetOntology}
\alias{IntelliGOComputeEiEj}
\alias{IntelliGOGetDepth}
\alias{IntelliGOGetParents}
\alias{IntelliGOGOSim}
\alias{IntelliGOInverseAnnotationFrequency}
\alias{rebuildICdata}
\alias{TCSSCompute_ICA}
\alias{TCSSComputeIC}
\alias{TCSSGetAncestors}
\alias{TCSSGetChildren}
\alias{TCSSGetOffsprings}
\alias{TCSSGoSim}
\alias{uniqsv}
\alias{WangGoSim}
\alias{WangMethod}
\alias{WangSemVal}
\alias{WangSemVal_internal}
\description{Internal ppiPre objects.}
\details{These are not to be called by the user.}
\keyword{internal}